Skip tests if required environment variables are not defined

+++ server/drivers/bigquery/test.js
@@ -22,10 +22,22 @@ describe('drivers/bigquery', function() {
   this.timeout(testTimeoutMsecs); // Set a large default timeout for all tests because BigQuery can be slow to respond.
 
   before(function() {
-    return bigquery
-      .runQuery(dropTable, connection)
-      .then(() => bigquery.runQuery(createTable, connection))
-      .then(() => bigquery.runQuery(inserts, connection));
+    if (
+      connection.projectId &&
+      connection.datasetName &&
+      connection.keyFile &&
+      connection.datasetLocation
+    ) {
+      bigquery
+        .runQuery(dropTable, connection)
+        .then(() => bigquery.runQuery(createTable, connection))
+        .then(() => bigquery.runQuery(inserts, connection));
+    } else {
+      console.log(
+        'Define BIGQUERY_TEST_GCP_PROJECT_ID, BIGQUERY_TEST_DATASET_NAME, BIGQUERY_TEST_CREDENTIALS_FILE, and BIGQUERY_TEST_DATASET_LOCATION to run the bigquery driver tests'
+      );
+      this.skip();
+    }
   });
 
   it('implements testConnection', function() {
@@ -146,27 +158,27 @@ describe('drivers/bigquery', function() {
         );
       });
   });
-});
 
-it('returns descriptive error message for bad syntax', function() {
-  let error;
+  it('returns descriptive error message for bad syntax', function() {
+    let error;
 
-  return bigquery
-    .runQuery(
-      `SELECT * FLARB ${connection.datasetName}.${testTable}`,
-      connection
-    )
-    .then(() => {
-      assert.fail('An error should have been thrown for bad syntax.');
-    })
-    .catch(e => {
-      error = e;
-    })
-    .then(() => {
-      assert(error, 'An error is thrown for bad syntax');
-      assert(
-        error.toString().includes('Syntax error'),
-        'throws the correct error'
-      );
-    });
+    return bigquery
+      .runQuery(
+        `SELECT * FLARB ${connection.datasetName}.${testTable}`,
+        connection
+      )
+      .then(() => {
+        assert.fail('An error should have been thrown for bad syntax.');
+      })
+      .catch(e => {
+        error = e;
+      })
+      .then(() => {
+        assert(error, 'An error is thrown for bad syntax');
+        assert(
+          error.toString().includes('Syntax error'),
+          'throws the correct error'
+        );
+      });
+  });
 });
